Add ability to page forward and backward through search results.  Similar functionality is coming for other paging objects.
Included a test that actually hits the API to test this.  This test will be the first in a set of tests that only run when the FULLTEST environment variable is set.  The full_tests.bat script facilitates running the full test suite on Windows (it just sets the environment and then forwards any args to the go test command).