Don't escape the query string before calling encode.
One of the features of the spotify API is allowing keyword filtering in the
query string. To search for a track called "holiday" by "akon" the user of the
API can supply the query string: "holiday artist:Akon". The escaped query
submitted to the server should be of the form: "holiday+artist%3AAkon". The
QueryEscape call will put the query string in this form, but when the Encode
function is called it escapes it once again resulting in
"holiday%2Bartist%253AAkon".