Fix ExternalURL collections
The current implementation seem to map these collections to a single
struct containing a Key/Value which in practice doesn't unmarshal to
anything. As dictated by the documentation [1], these are simple
objects mapping a string to a string and that's unlikely to change
during this API version, so I'd recommend using a simple map instead
(for user convenience).

[1] https://developer.spotify.com/web-api/get-playlist/