bugfix: does not get any audio features
fdc4ddb does not get any audio features from spotify. Updated
GetAudioFeatures to actually retrieve audio features based on the song
IDs. If an ID is not found it will now throw an error not place
a nil value in the appropriate position. Since we're returning
a slice of AudioFeatures pointers I believe it is too cumbersome for the
caller to have to check these afterwards.

TODO: Update tests as well.