Accept collaborative parameter for CreatePlaylistForUser
The collaborative parameter is optional in the Spotify Web API, and was
not supported in this library. I add the parameter to directly to the
CreatePlaylistForUser function, though a case could be made for writing
a separate function in order to avoid breaking existing calls.

Note that I do not check for the relationship between public and
collaborative in the body of the function, instead I let the Web API
return an error.

I added two tests for the private/public-collaborative permutations in
order to demonstrate the intended usage of the fields.