auth: remove contextwithHTTPClient
Setting a custom Transport with no ability for the caller to configure
it breaks the ability to configure an http proxy, while the default
behavior via http.DefaultTransport is to support it (via env),
see https://pkg.go.dev/net/http#RoundTripper.

This was added because Spotify didn't support HTTP/2 in 2017, but it
does now. Also, Go net/http should auto negotiate better.